refactor: rename file and remove 'utils' from name
